requirements.txt
opencv-python>=4.8.0
pyautogui>=0.9.54
Pillow>=10.0.0
pygetwindow>=0.0.9
pywin32>=306
numpy>=1.24.0
pygame>=2.5.0
